+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 * **Breaking**: Rename Kotlin factory methods to be named after Java interfaces instead of Java factory methods
+* **Breaking**: Rename `MessageEncoder/MessageDecoder.ofStrings()` to `ofString()`
 
 ## [0.3.0] - 2024-11-30